Config + Ordering
Added lengthened timeout for dask client + ability to allow failures and change timeout via command line. Had to move the order of the code/argparsing so that I can do this before anything is ever imported for dask. Also moved the setting of the dashboard to the main script as this does not appear to be working in jobs.py